--- a/test/mocha.opts
+++ b/test/mocha.opts
@@ -1,3 +1,2 @@
---ui mocha-typescript
---require source-map-support/register
-test/*.test.js
+--require ts-node/register
+test/*.ts