--- a/tsconfig.json
+++ b/tsconfig.json
@@ -2,6 +2,7 @@
   "compilerOptions": {
     "target": "es2015",
     "module": "commonjs",
+    "outDir": "dist",
     "sourceMap": true,
     "declaration": true,
     "noImplicitAny": true,
@@ -11,5 +12,10 @@
     "strictNullChecks": true,
     "suppressImplicitAnyIndexErrors": true
   },
+  "include": [
+    "src/*.ts",
+    "test/*.ts",
+    "test/benchmarks/*.ts"
+  ],
   "compileOnSave": true
 }