Also changes to using `dist/` as typescript's output directory to fix some build issues.

## Benchmarks

(v0: before, v1: after)

```
serializeIldcpResponse           v0 x 38,153 ops/sec ±0.84% (92 runs sampled)
serializeIldcpResponse           v1 x 407,086 ops/sec ±1.10% (91 runs sampled)
deserializeIldcpResponse         v0 x 311,105 ops/sec ±1.86% (89 runs sampled)
deserializeIldcpResponse         v1 x 1,440,833 ops/sec ±0.21% (92 runs sampled)
```